[#1] optimization in listeners
